refactor: 경로 수정 및 Trip 타입 통합
