--- src/types/trip.ts
@@ -1,7 +1,0 @@
-export interface Trip {
-  productId: number;
-  title: string;
-  start_date: string;
-  end_date: string;
-  price: number;
-}
